requirements.txt
azure-cognitiveservices-vision-customvision==3.1.0
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.15.0
azure-keyvault-secrets==4.8.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
Flask==3.0.2
fonttools==4.49.0
idna==3.6
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
msal==1.27.0
msal-extensions==1.1.0
msrest==0.7.1
numpy==1.26.4
oauthlib==3.2.2
packaging==23.2
pandas==2.2.1
pillow==10.2.0
portalocker==2.8.2
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
